[test] Factor PropertyUtils out of ChiselSpec
Move PropertyUtils into its own file.
